优化连接管理和心跳管理 去除失效的冗余部分

+++ Sora/Interfaces/ISoraConfig.cs
@@ -15,7 +15,7 @@ namespace Sora.Interfaces
         /// <summary>
         /// 服务器端口
         /// </summary>
-        uint Port { get; init; }
+        ushort Port { get; init; }
 
         /// <summary>
         /// 鉴权Token
